requirements.txt
pip==22.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-aliases==1.0.1
click-default-group==1.2.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-celery-beat==2.4.0
django-filter==22.1
django-rest==0.8.7
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
greenlet==2.0.1
idna==3.4
jsonpickle==3.0.1
kombu==5.2.4
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.4.0
prompt-toolkit==3.0.36
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-memcached==1.51
python3-openid==3.2.0
pytz==2022.7
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==9.13.0
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
tabulate==0.8.10
typing-extensions==3.10.0.2
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
djoser==2.1.0